+++ src/main/java/software/coolstuff/springframework/owncloud/service/api/OwncloudGroupQueryService.java
@@ -22,10 +22,10 @@ import java.util.List;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 
-import software.coolstuff.springframework.owncloud.model.OwncloudUserDetails;
+import software.coolstuff.springframework.owncloud.exception.auth.OwncloudGroupNotFoundException;
 
 /**
- * Get Information about Users from the Owncloud Server.
+ * Get Information about Groups from the Owncloud Server.
  * <p/>
  * This Service is only usable when the User has been authenticated
  * by the OwncloudAuthenticationProvider.
@@ -33,30 +33,39 @@ import software.coolstuff.springframework.owncloud.model.OwncloudUserDetails;
  * @author mufasa1976
  * @since 1.0.0
  */
-public interface OwncloudUserQueryService {
+public interface OwncloudGroupQueryService {
 
   /**
-   * Get the Details of a User
-   * @param username Name of the User
-   * @return Details of the User
-   * @throws AccessDeniedException Neither an Administrator nor the Owner
-   * @throws UsernameNotFoundException User doesn&apos;t exist
-   */
-  OwncloudUserDetails findOne(String username);
-
-  /**
-   * Find all Users.
-   * @return List of all Users
+   * Find all Groups.
+   * @return List of all Groups
    * @throws AccessDeniedException Not an Administrator
    */
   List<String> findAll();
 
   /**
-   * Find all Users whose Display Name confirm to the Filter Criterion.
-   * @param filter Filter Criterion for Display Name
-   * @return List of all Users whose Display Name confirm to the Filter Criterion
+   * Find all Groups whose Name confirm to the Filter Criterion.
+   * @param filter Filter Criterion for the Name
+   * @return List of all Groups whose Name confirm to the Filter Criterion
    * @throws AccessDeniedException Not an Administrator
    */
   List<String> findAll(String filter);
+
+  /**
+   * Find all Users that are members of a Group
+   * @param group Name of the Group
+   * @return List of all Users that are Members of the Group
+   * @throws AccessDeniedException Not an Administrator
+   * @throws OwncloudGroupNotFoundException Group doesn&apos;t exist
+   */
+  List<String> findAllUsers(String groupname);
+
+  /**
+   * Find all Groups of a User
+   * @param username Name of the User
+   * @return List of all Groups the User is a Member of
+   * @throws AccessDeniedException Neither an Administrator nor the Owner
+   * @throws UsernameNotFoundException User doesn&apos;t exist
+   */
+  List<String> findAllGroups(String username);
 
 }
