+++ src/main/java/software/coolstuff/springframework/owncloud/service/impl/rest/OwncloudRestUserQueryServiceImpl.java
@@ -24,10 +24,8 @@ import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.Validate;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.http.HttpMethod;
-import org.springframework.security.access.AccessDeniedException;
 
 import lombok.extern.slf4j.Slf4j;
-import software.coolstuff.springframework.owncloud.exception.auth.OwncloudGroupNotFoundException;
 import software.coolstuff.springframework.owncloud.model.OwncloudUserDetails;
 import software.coolstuff.springframework.owncloud.service.impl.rest.Ocs.User.Data.Quota;
 
@@ -39,12 +37,22 @@ class OwncloudRestUserQueryServiceImpl extends AbstractOwncloudRestServiceImpl i
   }
 
   @Override
-  public List<String> findAllUsers() {
-    return findAllUsers(null);
+  public OwncloudUserDetails findOne(String username) {
+    Validate.notBlank(username);
+    log.debug("Get Information about User {} from Location {}", username, getLocation());
+    Ocs.User user = exchange("/cloud/users/{user}", HttpMethod.GET, emptyEntity(), Ocs.User.class, username);
+    log.debug("Get all Groups assigned to User {} from Location {}", username, getLocation());
+    Ocs.Groups groups = exchange("/cloud/users/{user}/groups", HttpMethod.GET, emptyEntity(), Ocs.Groups.class, username);
+    return convert(username, user, groups);
   }
 
   @Override
-  public List<String> findAllUsers(String filter) {
+  public List<String> findAll() {
+    return findAll(null);
+  }
+
+  @Override
+  public List<String> findAll(String filter) {
     Ocs.Users users = null;
     if (StringUtils.isBlank(filter)) {
       log.debug("Get all Users by Filter Criteria {} from Location {}", filter, getLocation());
@@ -69,69 +77,6 @@ class OwncloudRestUserQueryServiceImpl extends AbstractOwncloudRestServiceImpl i
 
   private boolean isUsersNotNull(Ocs.Users ocsUsers) {
     return ocsUsers != null && ocsUsers.getData() != null && ocsUsers.getData().getUsers() != null;
-  }
-
-  @Override
-  public List<String> findAllGroups() {
-    return findAllGroups(null);
-  }
-
-  @Override
-  public List<String> findAllGroups(String filter) {
-    Ocs.Groups ocsGroups = null;
-    if (StringUtils.isBlank(filter)) {
-      log.debug("Get all Groups by Filter Criteria {} from Location {}", filter, getLocation());
-      ocsGroups = exchange("/cloud/groups", HttpMethod.GET, emptyEntity(), Ocs.Groups.class);
-    } else {
-      log.debug("Get all Groups from Location {}", getLocation());
-      ocsGroups = exchange("/cloud/groups?search={filter}", HttpMethod.GET, emptyEntity(), Ocs.Groups.class, filter);
-    }
-    return OwncloudRestUtils.convertGroups(ocsGroups);
-  }
-
-  @Override
-  public List<String> findAllMembersOfGroup(String groupname) {
-    Validate.notBlank(groupname);
-    log.debug("Get all Users assigned to Group {} from Location {}", groupname, getLocation());
-    Ocs.Users users = exchange("/cloud/groups/{group}", HttpMethod.GET, emptyEntity(), Ocs.Users.class, (authorizationUser, uri, meta) -> {
-      if ("ok".equals(meta.getStatus())) {
-        return;
-      }
-
-      String exceptionMessage;
-      switch (meta.getStatuscode()) {
-        case 997:
-          exceptionMessage = String.format("User %s is not authorized to access Resource %s", authorizationUser, uri);
-          log.warn("Error 997: {}", exceptionMessage);
-          throw new AccessDeniedException(exceptionMessage);
-        case 998:
-          log.error("Error 998: Group {} not found", groupname);
-          throw new OwncloudGroupNotFoundException(groupname);
-        default:
-          exceptionMessage = String.format("Unknown Error Code %d. Reason: %s", meta.getStatuscode(), StringUtils.defaultIfEmpty(meta.getMessage(), ""));
-          log.error(exceptionMessage);
-          throw new IllegalStateException(exceptionMessage);
-      }
-    }, groupname);
-    return convertUsers(users);
-  }
-
-  @Override
-  public List<String> findAllGroupsOfUser(String username) {
-    Validate.notBlank(username);
-    log.debug("Get all Groups assigned to User {} from Location {}", username, getLocation());
-    Ocs.Groups ocsGroups = exchange("/cloud/users/{user}/groups", HttpMethod.GET, emptyEntity(), Ocs.Groups.class, username);
-    return OwncloudRestUtils.convertGroups(ocsGroups);
-  }
-
-  @Override
-  public OwncloudUserDetails findOneUser(String username) {
-    Validate.notBlank(username);
-    log.debug("Get Information about User {} from Location {}", username, getLocation());
-    Ocs.User user = exchange("/cloud/users/{user}", HttpMethod.GET, emptyEntity(), Ocs.User.class, username);
-    log.debug("Get all Groups assigned to User {} from Location {}", username, getLocation());
-    Ocs.Groups groups = exchange("/cloud/users/{user}/groups", HttpMethod.GET, emptyEntity(), Ocs.Groups.class, username);
-    return convert(username, user, groups);
   }
 
   @Override
