+++ src/test/java/software/coolstuff/springframework/owncloud/service/impl/local/AbstractLocalOwncloudUserModificationServiceTest.java
@@ -1,7 +1,7 @@
 package software.coolstuff.springframework.owncloud.service.impl.local;
 
-import software.coolstuff.springframework.owncloud.service.AbstractOwncloudUserModificationServiceTest;
+import software.coolstuff.springframework.owncloud.service.AbstractOwncloudUserServiceTest;
 
-public abstract class AbstractLocalOwncloudUserModificationServiceTest extends AbstractOwncloudUserModificationServiceTest {
+public abstract class AbstractLocalOwncloudUserModificationServiceTest extends AbstractOwncloudUserServiceTest {
 
 }
