+++ src/test/java/software/coolstuff/springframework/owncloud/service/impl/local/OwncloudLocalGroupServiceNoModificationTest.java
@@ -15,7 +15,7 @@
    along with this program; if not, write to the Free Software Foundation,
    Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301  USA
 */
-package software.coolstuff.springframework.owncloud.service.impl.rest;
+package software.coolstuff.springframework.owncloud.service.impl.local;
 
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,32 +24,27 @@ import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.context.ActiveProfiles;
 
-import software.coolstuff.springframework.owncloud.model.OwncloudModificationUser;
+import software.coolstuff.springframework.owncloud.service.api.OwncloudGroupService;
 import software.coolstuff.springframework.owncloud.service.api.OwncloudUserService;
 import software.coolstuff.springframework.owncloud.service.impl.AbstractOwncloudServiceTest;
 
 @RestClientTest(OwncloudUserService.class)
-@ActiveProfiles("REST-NO-MODIFICATION-USER-SERVICE")
-public class OwncloudRestUserModificationServiceNoModificationTest extends AbstractOwncloudServiceTest implements OwncloudRestServiceTest {
+@ActiveProfiles("LOCAL-NO-MODIFICATION-USER-SERVICE")
+public class OwncloudLocalGroupServiceNoModificationTest extends AbstractOwncloudServiceTest {
 
   @Autowired
-  private OwncloudUserService userModificationService;
+  private OwncloudGroupService groupService;
 
-  @Override
-  public final OwncloudRestService owncloudService() {
-    return (OwncloudRestService) userModificationService;
+  @Test(expected = AccessDeniedException.class)
+  @WithMockUser(username = "user", password = "s3cr3t")
+  public void testCreateGroup() {
+    groupService.create("shouldBeAccessDenied");
   }
 
   @Test(expected = AccessDeniedException.class)
   @WithMockUser(username = "user", password = "s3cr3t")
-  public void testSaveUser() {
-    userModificationService.save(new OwncloudModificationUser("user99"));
-  }
-
-  @Test(expected = AccessDeniedException.class)
-  @WithMockUser(username = "user", password = "s3cr3t")
-  public void testDeleteUser() {
-    userModificationService.delete("shouldBeAccessDenied");
+  public void testDeleteGroup() {
+    groupService.delete("shouldBeAccessDenied");
   }
 
 }
