+++ src/test/java/software/coolstuff/springframework/owncloud/service/impl/local/OwncloudLocalUserServiceNoModificationTest.java
@@ -15,7 +15,7 @@
    along with this program; if not, write to the Free Software Foundation,
    Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301  USA
 */
-package software.coolstuff.springframework.owncloud.service.impl.rest;
+package software.coolstuff.springframework.owncloud.service.impl.local;
 
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -29,27 +29,22 @@ import software.coolstuff.springframework.owncloud.service.api.OwncloudUserServi
 import software.coolstuff.springframework.owncloud.service.impl.AbstractOwncloudServiceTest;
 
 @RestClientTest(OwncloudUserService.class)
-@ActiveProfiles("REST-NO-MODIFICATION-USER-SERVICE")
-public class OwncloudRestUserModificationServiceNoModificationTest extends AbstractOwncloudServiceTest implements OwncloudRestServiceTest {
+@ActiveProfiles("LOCAL-NO-MODIFICATION-USER-SERVICE")
+public class OwncloudLocalUserServiceNoModificationTest extends AbstractOwncloudServiceTest {
 
   @Autowired
-  private OwncloudUserService userModificationService;
-
-  @Override
-  public final OwncloudRestService owncloudService() {
-    return (OwncloudRestService) userModificationService;
-  }
+  private OwncloudUserService userService;
 
   @Test(expected = AccessDeniedException.class)
   @WithMockUser(username = "user", password = "s3cr3t")
   public void testSaveUser() {
-    userModificationService.save(new OwncloudModificationUser("user99"));
+    userService.save(new OwncloudModificationUser("user99"));
   }
 
   @Test(expected = AccessDeniedException.class)
   @WithMockUser(username = "user", password = "s3cr3t")
   public void testDeleteUser() {
-    userModificationService.delete("shouldBeAccessDenied");
+    userService.delete("shouldBeAccessDenied");
   }
 
 }
